chore: add metrics and reduce bloat by moving in dedicated package
